renamed goaldb to goaldatabase
